Fix namespaces after renaming folder
