+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_021/ForwardMeasurements/V1/Model/ForwardMeteredDataAcceptedV1.cs
@@ -15,7 +15,7 @@
 using Energinet.DataHub.ProcessManager.Components.Abstractions.EnqueueActorMessages;
 using Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData.V1.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeasurements.V1.Model;
 
 /// <summary>
 /// A model containing the data required for notifying market actors when new metered data has been accepted.
