+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_021/ForwardMeasurements/V1/Model/ForwardMeteredDataInputV1.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData.V1.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeasurements.V1.Model;
 
 /// <summary>
 /// An immutable input to start the orchestration instance for "BRS_021_ForwardMeteredData" V1.
