+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_021/ForwardMeasurements/V1/Model/ForwardMeteredDataNotifyEventV1.cs
@@ -15,7 +15,7 @@
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 using Energinet.DataHub.ProcessManager.Abstractions.Client;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData.V1.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeasurements.V1.Model;
 
 public record ForwardMeteredDataNotifyEventV1(
     string OrchestrationInstanceId)
