+++ source/ProcessManager.Orchestrations.Abstractions/Processes/BRS_021/ForwardMeasurements/V1/Model/ForwardMeteredDataRejectedV1.cs
@@ -17,7 +17,7 @@ using Energinet.DataHub.ProcessManager.Components.Abstractions.BusinessValidatio
 using Energinet.DataHub.ProcessManager.Components.Abstractions.EnqueueActorMessages;
 using Energinet.DataHub.ProcessManager.Components.Abstractions.ValueObjects;
 
-namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeteredData.V1.Model;
+namespace Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_021.ForwardMeasurements.V1.Model;
 
 public record ForwardMeteredDataRejectedV1(
     string OriginalActorMessageId,
